requirements.txt
python-telegram-bot
python-dotenv
flask
supabase
